reorganized sidebar and privatenav
