+++ client/App.js
@@ -3,7 +3,7 @@ import { Routes, Route, Outlet } from "react-router-dom";
 
 import ".././client/index.css";
 
-import Navigation from "./components/Navigation.js";
+import Navigation from "./components/Navigation.jsx";
 import Login from "./subpage/Login.js";
 import Home from "./subpage/Home.js";
 import About from "./subpage/About.js";
@@ -13,7 +13,9 @@ import Contact from "./subpage/Contact.js";
 import Signup from "./subpage/Signup.js";
 import Profile from "./subpage/Profile.js";
 import Unauthorized from "./subpage/Unauthorized.js";
-import RequiredAuth from "./components/RequiredAuth.js";
+import RequiredAuth from "./components/RequiredAuth.jsx";
+import ProjectList from "./subpage/ProjectList.js";
+import PrivateNav from "./components/PrivateNav.jsx";
 
 const RoutesWithNavigation = () => {
   return (
@@ -48,16 +50,19 @@ const App = () => {
           <Route path="/unauthorized" element={<Unauthorized />} />
         </Route>
 
-        <Route element={<RequiredAuth />}>
-          <Route path="/profile" element={<Profile />} />
+        <Route path="/profile" element={<RequiredAuth />}>
+          <Route element={<PrivateNav />} >
+            <Route index element={<Profile />} />
+            <Route path="/projectlist" element={<ProjectList />} />
+          </Route>
         </Route>
 
-        {/* <Route path="/profile/basic/:id" element={
+        {/* <Route path="/profile/project" element={
             <RequiredAuth>
               <Profile />
             </RequiredAuth>
-        } />
-        <Route path="/profile/admin/:id" element={
+        } /> */}
+        {/* <Route path="/profile/admin/:id" element={
             <RequiredAuth>
               <Profile />
             </RequiredAuth>
