+++ client/components/ProjectCards.jsx
@@ -1,0 +1,10 @@
+import React, { useState, useEffect } from "react";
+import { Link, useNavigate, navLink } from "react-router-dom";
+import { FontAwesomeIcon as FAIcon } from "@fortawesome/react-fontawesome";
+import { faQuestionCircle } from "@fortawesome/free-regular-svg-icons";
+
+const ProjectCards = (props) => {
+    
+}
+
+export default ProjectCards;
